Circumvent issue of Rollup not inlining import of image.js
